ALLJOYN-2100 Increase stack size limit for Java VM compatibility
The previous stack size of 80k resulted in an error from
AttachCurrentThread when attempting to call in to a Java 7 JVM (Java
1.7.0_25).

Increasing the stack size to 100k results in a success return value
from AttachCurrentThread, but Java raises a stack overflow exception
soon after.

128k gives no error in limited testing, but this patch sets the stack
limit to 256k to give some headroom.

Change-Id: Idd25f89fb624a845990b6667519227490ac75363