ALLJOYN-2210 Fixed compiler warning due to our definition of NaN
When compiling code with /W2 flag our definition of NAN would return 2 warnings
   - "truncation from __int64 to const unsigned long"
   - "truncation of constant value"

The original number given was 36-bits long when it should have been 32-bits

Looking up the IEEE-754 definition for a quiet NaN we should be using
 0x7FFFFFFF (quiet NaN) or 0xFFFFFFFF (negative quiet NaN) both these numbers
are 32-bits in length.  I made the assumption that the original intent of this
code was to have a 32-bit quiet NaN.
converted to a float that is 32-bits.

Change-Id: I23f19afba4ea68394790223479d91d739b397d9a